+++ src/components/cursor-trail/CursorTrail.css
