+++ src/components/cursor-trail/CursorTrail.tsx
